backend/shared-infrastructure/flink-processing/requirements_mimic.txt
# MIMIC-IV Model Training Requirements
# Install with: pip install -r requirements_mimic.txt

# GCP BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Model export
onnx>=1.15.0
onnxruntime>=1.17.0
onnxmltools>=1.12.0
skl2onnx>=1.16.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
